pyproject.toml
[project]
name = "local-talking-llm"
version = "2.0.0"
description = "Voice-controlled personal assistant with orchestrated model routing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.0.0",
    "httpx>=0.28.0",
    "langchain-ollama>=0.3.3",
    "langchain>=0.3.25",
    "openai-whisper>=20240930",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "piper-tts>=1.4.0",
    "PyYAML>=6.0.2",
    "rich>=14.0.0",
    "sentence-transformers>=3.0.0",
    "sounddevice>=0.5.2",
    "torch>=2.0.0",
    "webrtcvad-wheels>=2.0.14",
    "zvec>=0.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.9.0",
]
